UHF-6540: Test menu storage
