+++ public/modules/custom/helfi_global_navigation/tests/src/Kernel/GlobalMenuResourceTest.php
@@ -5,8 +5,6 @@ declare(strict_types = 1);
 namespace Drupal\Tests\helfi_global_navigation\Kernel;
 
 use Drupal\helfi_global_navigation\Entity\GlobalMenu;
-use Drupal\language\Entity\ConfigurableLanguage;
-use Drupal\Tests\helfi_api_base\Kernel\ApiKernelTestBase;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 /**
@@ -14,42 +12,7 @@ use Symfony\Component\HttpFoundation\Response as HttpResponse;
  *
  * @group helfi_global_navigation
  */
-class GlobalMenuResourceTest extends ApiKernelTestBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  protected static $modules = [
-    'user',
-    'serialization',
-    'migrate',
-    'entity',
-    'rest',
-    'language',
-    'json_field',
-    'content_translation',
-    'helfi_global_navigation',
-  ];
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setUp(): void {
-    parent::setUp();
-
-    $this->installEntitySchema('global_menu');
-    $this->installConfig('helfi_global_navigation');
-    $this->installConfig(['language', 'content_translation']);
-
-    foreach (['fi', 'sv'] as $langcode) {
-      ConfigurableLanguage::createFromLangcode($langcode)->save();
-    }
-    $this->config('language.negotiation')
-      ->set('url.prefixes', ['en' => 'en', 'fi' => 'fi', 'sv' => 'sv'])
-      ->save();
-
-    \Drupal::service('kernel')->rebuildContainer();
-  }
+class GlobalMenuResourceTest extends KernelTestBase {
 
   /**
    * Gets the mocked JSON.
